move: Interface 및 Implementation 폴더 이동 (#27)
