+++ Tests/Macros/Deprecated/DeprecatedMacroTests.swift
